Update server address and project name
Changed the server address in the environment configuration file from
"http://127.0.0.1" to "http://10.10.73.40". Also updated the project
name in the package.json file from "sport-portal" to
"sport-portal-backend". These changes were made to ensure the backend
server is connecting to the correct address and to maintain consistency
in project naming conventions. (\#42)